add message deletion

+++ app/bot/register/handlers.py
@@ -6,6 +6,7 @@ from aiogram.methods import SendMessage
 from aiogram.types import Message, CallbackQuery
 from asyncpg import Record
 
+from app import main_backend
 from app.bot import crud, keyboards
 from app.bot.register.states import Register, Location
 
@@ -21,14 +22,12 @@ async def start(message: Message, state: FSMContext):
 
     await state.set_state(Register.password)
 
-    # delete_list: list[Message | SendMessage] = \
-    #    await add_messages_to_delete_list(user_message=message,
-    # bot_messages=messages)
-
+    delete_list = [message]
     for msg in messages:
-        await message.answer(msg['text'])
+        m = await message.answer(msg['text'])
+        delete_list.append(m)
 
-    await state.update_data(count=0)
+    await state.update_data(count=0, delete_list=delete_list, locations=[])
 
 
 @router.message(F.text, Register.password)
@@ -50,6 +49,8 @@ async def enter_password(message: Message, state: FSMContext):
                             messages=messages, state=state):
         return
 
+    data = await state.get_data()
+    data['delete_list'].append(message)
     for num, msg in enumerate(messages['correct']):
         kb = None
         if msg['id'] in inline_kbs:
@@ -57,16 +58,18 @@ async def enter_password(message: Message, state: FSMContext):
             kb = keyboards.create_inline_keyboard(buttons=buttons)
 
         m = await message.answer(msg['text'], reply_markup=kb)
-        # data['delete_list'].append(m)
-    # data['delete_list'].append(message)
+        data['delete_list'].append(m)
+
+    await state.update_data(delete_list=data['delete_list'])
+    await delete_messages(state, count=len(messages['correct']))
     await state.set_state(Register.choose_option)
 
 
 @router.callback_query(Register.choose_option, F.data == 'referral code')
 async def begin_enter_referral_code(callback: CallbackQuery,
                                     state: FSMContext):
-    # data = await state.get_data()
-
+    data = await state.get_data()
+    print(data.keys())
     state_name: str = await state.get_state()
     state_name += ':referral_code'
     messages, _, reply_kbs = \
@@ -76,11 +79,11 @@ async def begin_enter_referral_code(callback: CallbackQuery,
 
     for msg in messages:
         m = await callback.message.answer(msg['text'])
-        # data['delete_list'].append(m)
+        data['delete_list'].append(m)
 
-    # print(222222)
-    # print([x.text for x in data['delete_list']])
-    # await state.update_data(delete_list=data['delete_list'])
+    print(222222)
+    print([x.text for x in data['delete_list']])
+    await state.update_data(delete_list=data['delete_list'])
     await state.set_state(Register.referral_code)
     await callback.answer()
 
@@ -101,9 +104,11 @@ async def enter_referral_code(message: Message, state: FSMContext):
         keys=['incorrect', 'used', 'correct'])
 
     if await check_referral_code(code=code, messages=messages,
-                                 message=message):
+                                 message=message, state=state):
         return
 
+    data = await state.get_data()
+    data['delete_list'].append(message)
     for num, msg in enumerate(messages['correct']):
         kb = None
         if msg['id'] in inline_kbs:
@@ -111,17 +116,17 @@ async def enter_referral_code(message: Message, state: FSMContext):
             kb = keyboards.create_inline_keyboard(buttons=buttons)
 
         m = await message.answer(msg['text'], reply_markup=kb)
-        # data['delete_list'].append(m)
-    # data['delete_list'].append(message)
-    # print(333333)
-    # print([x.text for x in data['delete_list']])
+        data['delete_list'].append(m)
+
+    print(333333)
+    print([x.text for x in data['delete_list']])
     # await state.update_data(delete_list=data['delete_list'])
+    await delete_messages(state, len(messages['correct']))
     await state.clear()
 
 
 @router.callback_query(Register.choose_option, F.data == 'location')
 async def begin_enter_locality(callback: CallbackQuery, state: FSMContext):
-    await state.set_data({})
     await state.set_state(Location.country)
     state_name = 'Location:start'
 
@@ -130,24 +135,48 @@ async def begin_enter_locality(callback: CallbackQuery, state: FSMContext):
                                       with_inline_kb=False,
                                       with_reply_kb=True)
 
-    # async with aiohttp.ClientSession() as session:
-    #    async with session.get(
-    #            'https://iqpr.cc/api/v1/locality_list/COUNTRY') as response:
-    #        response = await response.json()
-    #
-    # data = {}
-    # for country in response['data']:
-    #    data[country['id']] = country['title']
+    locations = await main_backend.get_location_list(location="COUNTRY")
+    data = {}
 
-    for num, msg in enumerate(messages):
+    print(messages)
+
+    messages: dict[str, list] = get_message_cases(
+        messages=messages,
+        keys=['inline', 'base'])
+
+    print(messages)
+    for country in locations['data']:
+        data[str(country['id'])] = country['title']
+
+    print(data)
+
+    inline_kb = keyboards.create_inline_keyboard(buttons=data.values(),
+                                                 callbacks=data.keys())
+
+    print(inline_kb)
+    data = await state.get_data()
+    for num, msg in enumerate(messages['inline']):
+        kb = None
+        if num == len(messages['inline']) - 1:
+            kb = inline_kb
+
+        m = await callback.message.answer(msg['text'], reply_markup=kb)
+        data['delete_list'].append(m)
+
+    for num, msg in enumerate(messages['base']):
+
         kb = None
 
         if num == 0:
             buttons: list[str] = [x['text'] for x in reply_buttons]
             kb = keyboards.create_reply_keyboard(buttons=buttons)
 
-        await callback.message.answer(msg['text'], reply_markup=kb)
-        await callback.answer()
+        m = await callback.message.answer(msg['text'], reply_markup=kb)
+        data['delete_list'].append(m)
+    await state.update_data(locations=locations,
+                            delete_list=data['delete_list'])
+
+    await callback.answer()
 
 
 class ButtonFilter(Filter):
@@ -173,6 +202,7 @@ async def _save(message: Message, state: FSMContext):
             data=data)
         print(response.status)
 
+    await delete_messages(state, 1)
     await state.clear()
 
 
@@ -180,58 +210,142 @@ async def _save(message: Message, state: FSMContext):
 async def enter_country(message: Message, state: FSMContext):
     country = message.text
     await state.update_data(country=country)
+    data = await state.get_data()
 
-    state_name = await state.get_state()
-    if not await check(state_name=state_name,
-                       message=message,
-                       location='COUNTRY'):
+    if not await check(message=message, state=state):
         await state.set_state(Location.federal_region)
+
+
+@router.callback_query(Location.country)
+async def enter_country_callback(callback: CallbackQuery, state: FSMContext):
+    country_id = callback.data
+    await state.update_data(country_id=country_id)
+    await check(message=callback.message,
+                state=state,
+                callback_data=callback.data)
+    await state.set_state(Location.federal_region)
+    await callback.answer()
 
 
 @router.message(Location.federal_region, F.text)
 async def enter_federal_region(message: Message, state: FSMContext):
     federal_region = message.text
     await state.update_data(federal_region=federal_region)
-
-    state_name = await state.get_state()
-
-    if not await check(state_name=state_name,
-                       message=message,
-                       location='FEDERAL_REGION'):
+    if not await check(message=message, state=state):
         await state.set_state(Location.region)
+
+
+@router.callback_query(Location.federal_region)
+async def enter_federal_region_callback(callback: CallbackQuery,
+                                        state: FSMContext):
+    federal_region_id = callback.data
+    await state.update_data(federal_region_id=federal_region_id)
+    await check(message=callback.message,
+                state=state,
+                callback_data=callback.data)
+    await state.set_state(Location.region)
+    await callback.answer()
 
 
 @router.message(Location.region, F.text)
 async def enter_region(message: Message, state: FSMContext):
     region = message.text
     await state.update_data(region=region)
-
-    state_name = await state.get_state()
-
-    if not await check(state_name=state_name,
-                       message=message,
-                       location='REGION'):
+    if not await check(message=message, state=state):
         await state.set_state(Location.city)
+
+
+@router.callback_query(Location.region)
+async def enter_region_callback(callback: CallbackQuery, state: FSMContext):
+    region_id = callback.data
+    await state.update_data(region_id=region_id)
+    await check(message=callback.message,
+                state=state,
+                callback_data=callback.data)
+    await state.set_state(Location.city)
+    await callback.answer()
 
 
 @router.message(Location.city, F.text)
 async def enter_city(message: Message, state: FSMContext):
     city = message.text
+    data = await state.get_data()
+
     await state.update_data(city=city)
+    state_name = await state.get_state()
+
+    messages, _, __ = \
+        await crud.get_state_messages(state_name=state_name,
+                                      with_inline_kb=False,
+                                      with_reply_kb=False)
+    messages: dict[str, list] = get_message_cases(
+        messages=messages,
+        keys=['not_found', 'inline', 'correct'])
+
+    state_data = await state.get_data()
+    locations = state_data['locations']
+
+    titles = []
+
+    for location in locations['data']:
+        titles.append(location['title'])
+
+    data['delete_list'].append(message)
+    if message.text not in titles:
+        for msg in messages['not_found']:
+            m = await message.answer(msg['text'])
+            data['delete_list'].append(m)
+
+        return True
+
+    for msg in messages['inline']:
+        m = await message.answer(msg['text'])
+        data['delete_list'].append(m)
+
+    await state.update_data(delete_list=data['delete_list'])
+    await state.set_state(Location.district)
+
+
+@router.callback_query(Location.city)
+async def enter_city_callback(callback: CallbackQuery, state: FSMContext):
+    city_id = callback.data
+    await state.update_data(city_id=city_id)
+    data = await state.get_data()
 
     state_name = await state.get_state()
 
-    if not await check(state_name=state_name,
-                       message=message,
-                       location='CITY'):
-        await state.set_state(Location.district)
+    messages, _, __ = \
+        await crud.get_state_messages(state_name=state_name,
+                                      with_inline_kb=False,
+                                      with_reply_kb=False)
+    messages: dict[str, list] = get_message_cases(
+        messages=messages,
+        keys=['not_found', 'inline', 'correct'])
+
+    state_data = await state.get_data()
+    locations = state_data['locations']
+
+    titles = []
+
+    for location in locations['data']:
+        titles.append(location['title'])
+
+    for msg in messages['inline']:
+        m = await callback.message.answer(msg['text'])
+        data['delete_list'].append(m)
+
+    await state.update_data(delete_list=data['delete_list'])
+    await state.set_state(Location.district)
+    data = await state.get_data()
+    print(111111111, data)
+    await callback.answer()
 
 
 @router.message(Location.district, F.text)
 async def enter_district(message: Message, state: FSMContext):
     district = message.text
     await state.update_data(district=district)
-
+    data = await state.get_data()
     state_name = await state.get_state()
 
     messages, _, __ = \
@@ -241,32 +355,40 @@ async def enter_district(message: Message, state: FSMContext):
 
     messages: dict[str, list] = get_message_cases(
         messages=messages,
-        keys=['not_found', 'correct'])
+        keys=['inline', 'correct'])
 
-    for msg in messages['correct']:
-        await message.answer(msg['text'])
+    data['delete_list'].append(message)
+    for msg in messages['inline']:
+        m = await message.answer(msg['text'])
+        data['delete_list'].append(m)
 
+    await state.update_data(delete_list=data['delete_list'])
     await state.set_state(Location.micro_district)
 
 
 @router.message(Location.micro_district, F.text)
 async def enter_micro_district(message: Message, state: FSMContext):
     micro_district = message.text
-
+    data = await state.get_data()
     await state.update_data(micro_district=micro_district)
 
     state_name = await state.get_state()
     state_name += ':success'
 
-    messages, _, __ = \
+    messages, inline_kbs, __ = \
         await crud.get_state_messages(state_name=state_name,
-                                      with_inline_kb=False,
+                                      with_inline_kb=True,
                                       with_reply_kb=False)
 
+    data['delete_list'].append(message)
     for msg in messages:
-        await message.answer(msg['text'])
-
-    # await message.answer(str(await state.get_data()))
+        kb = None
+        if msg['id'] in inline_kbs:
+            buttons: list[str] = [x['text'] for x in inline_kbs[msg['id']]]
+            kb = keyboards.create_inline_keyboard(buttons=buttons)
+        m = await message.answer(msg['text'], reply_markup=kb)
+        data['delete_list'].append(m)
+    await state.update_data(delete_list=data['delete_list'])
     await _save(message=message, state=state)
 
 
@@ -289,15 +411,27 @@ async def add_messages_to_delete_list(
     return delete_list
 
 
-async def delete_messages(state: FSMContext):
+async def delete_messages(state: FSMContext, count: int):
     data = await state.get_data()
     messages: list[Message | SendMessage] = data['delete_list']
-    print([x.text for x in messages])
-    for num, m in enumerate(messages):
-        # if num <= len(messages) - 1:
-        await m.delete()
 
-    await state.set_data({})
+    print(count)
+    print(len(messages))
+    print([x.text for x in messages])
+    messages_to_remove = []
+    for num, m in enumerate(messages.copy()):
+
+        if num + count + 1 > len(messages):
+            break
+        print(m.text)
+        await m.delete()
+        messages_to_remove.append(m)
+
+    for m in messages_to_remove:
+        messages.remove(m)
+
+    print(10101010110101010101, [x.text for x in messages])
+    await state.update_data(delete_list=messages)
 
 
 def get_message_cases(messages: list[Record],
@@ -325,42 +459,63 @@ async def check_password(password: str,
                          state: FSMContext,
                          messages: dict,
                          message: Message) -> int:
+    state_data = await state.get_data()
+    delete_list = state_data['delete_list']
+
     if password != '1234':
+        delete_list.append(message)
         data = await state.get_data()
         count = data['count'] + 1
 
         if count == 3:
 
             for msg in messages['incorrect_3_times']:
-                await message.answer(msg['text'])
-            await state.update_data(count=0)
+                m = await message.answer(msg['text'])
+                delete_list.append(m)
+
+            await state.update_data(count=0, delete_list=delete_list)
             await state.set_state(None)
+            data = await state.get_data()
+            delete_list = data['delete_list']
+
+            await delete_messages(state, 1)
             return True
 
         for msg in messages['incorrect']:
-            await message.answer(msg['text'])
-            await state.update_data(count=count)
+            m = await message.answer(msg['text'])
+            delete_list.append(m)
+            await state.update_data(count=count, delete_list=delete_list)
         return True
 
 
 async def check_referral_code(code: str,
                               messages: dict,
+                              state: FSMContext,
                               message: Message) -> bool:
+    state_data = await state.get_data()
+    delete_list = state_data['delete_list']
     if code != '5678':
-
+        delete_list.append(message)
         if code == '5555':
             for msg in messages['used']:
-                await message.answer(msg['text'])
+                m = await message.answer(msg['text'])
+                delete_list.append(m)
+            print('aaaaaaaaaaaaaaaaaaaaaaaaaa', [x.text for x in delete_list])
+            await state.update_data(delete_list=delete_list)
             return True
 
         for msg in messages['incorrect']:
-            await message.answer(msg['text'])
+            m = await message.answer(msg['text'])
+            delete_list.append(m)
+        print('bbbbb', [x.text for x in delete_list])
+        await state.update_data(delete_list=delete_list)
         return True
 
 
-async def check(state_name: str,
-                message: Message,
-                location: str):
+async def check(message: Message, state: FSMContext,
+                callback_data=None):
+    state_name = await state.get_state()
+
     messages, _, __ = \
         await crud.get_state_messages(state_name=state_name,
                                       with_inline_kb=False,
@@ -368,19 +523,52 @@ async def check(state_name: str,
 
     messages: dict[str, list] = get_message_cases(
         messages=messages,
-        keys=['not_found', 'correct'])
+        keys=['not_found', 'inline', 'correct'])
 
-    async with aiohttp.ClientSession() as session:
-        async with session.get(
-                f'https://iqpr.cc/api/v1/locality_list/{location}') as response:
-            response = await response.json()
+    state_data = await state.get_data()
+    locations = state_data['locations']
 
-    titles = [x['title'] for x in response['data']]
+    titles = []
+    ids = []
+    for location in locations['data']:
+        titles.append(location['title'])
+        ids.append(location['id'])
+    location = locations['data'][0]['locality']
 
-    if message.text not in titles:
-        for msg in messages['not_found']:
-            await message.answer(msg['text'])
-        return True
+    if not callback_data:
+        state_data['delete_list'].append(message)
+        if message.text not in titles:
+            for msg in messages['not_found']:
+                m = await message.answer(msg['text'])
+                state_data['delete_list'].append(m)
+            return True
 
-    for msg in messages['correct']:
-        await message.answer(msg['text'])
+        index = titles.index(message.text)
+        parent_id = ids[index]
+    else:
+        parent_id = callback_data
+    locations = await main_backend.get_locations_with_parent(
+        location=location,
+        parent_id=parent_id)
+    data = {}
+
+    for location in locations['data']:
+        data[str(location['id'])] = location['title']
+
+    print(data)
+
+    inline_kb = keyboards.create_inline_keyboard(buttons=data.values(),
+                                                 callbacks=data.keys())
+
+    for num, msg in enumerate(messages['inline']):
+        kb = None
+        if num == len(messages['inline']) - 1:
+            kb = inline_kb
+
+        m = await message.answer(msg['text'], reply_markup=kb)
+        state_data['delete_list'].append(m)
+    for num, msg in enumerate(messages['correct']):
+        m = await message.answer(msg['text'])
+        state_data['delete_list'].append(m)
+    await state.update_data(locations=locations,
+                            delete_list=state_data['delete_list'])
